# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cloe API Documentation",
        "description": "Cloe API documentation",
        "contact": {
            "name": "Jesus Romero",
            "email": "gisuss8520@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://cloeBE.test",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/centro-acopio/index": {
            "get": {
                "tags": [
                    "Centro Acopio"
                ],
                "summary": "INDEX Centro Acopio",
                "description": "Retorna la lista de Centro Acopio registrados en el sistema",
                "operationId": "607b6ca82685a35b671231db8014c3df",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/centro-acopio/store": {
            "post": {
                "tags": [
                    "Centro Acopio"
                ],
                "summary": "Crea un nuevo Centro de Acopio",
                "description": "Registra un nuevo Centro de Acopio.",
                "operationId": "f5d29c143d19a4b8ea428534f12be6d9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCentroRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso creado con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/centro-acopio/update/{centro_id}": {
            "put": {
                "tags": [
                    "Centro Acopio"
                ],
                "summary": "Actualiza datos de un Centro de Acopio",
                "description": "Actualiza datos de un Centro de Acopio.",
                "operationId": "95c9b797650aaa57800a56a8bd8979bc",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCentroRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso actualizado con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/centro-acopio/show/{centro_id}": {
            "get": {
                "tags": [
                    "Centro Acopio"
                ],
                "summary": "SHOW Centro de Acopio",
                "description": "Retorna la información del Centro de Acopio solicitado.",
                "operationId": "6fe59c0517191e63a2f4f95b7b9cc9e4",
                "parameters": [
                    {
                        "name": "centro_id",
                        "in": "path",
                        "description": "Parámetro necesario para la consulta en la tabla Centros de Acopio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        },
                        "examples": {
                            "integer": {
                                "summary": "Introduce el número de id de un centro de acopio.",
                                "value": 1
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "No se ha encontrado el centro de acopio.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/centro-acopio/activate/{centro_id}": {
            "post": {
                "tags": [
                    "Centro Acopio"
                ],
                "summary": "ACTIVAR Centro de Acopio",
                "description": "Reactiva el Centro de Acopio solicitado.",
                "operationId": "faec7e71d2a4b963b4d5024da0754824",
                "parameters": [
                    {
                        "name": "centro_id",
                        "in": "path",
                        "description": "Parámetro necesario para la consulta en la tabla Centros de Acopio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        },
                        "examples": {
                            "integer": {
                                "summary": "Introduce el número de id de un centro de acopio.",
                                "value": 1
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso actualizado con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "No se ha encontrado el centro de acopio.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/centro-acopio/desactivate/{centro_id}": {
            "post": {
                "tags": [
                    "Centro Acopio"
                ],
                "summary": "DESACTIVAR Centro de Acopio",
                "description": "Desactiva el Centro de Acopio solicitado.",
                "operationId": "9a54a9abd2e52cffd8ecdb74ff9381c1",
                "parameters": [
                    {
                        "name": "centro_id",
                        "in": "path",
                        "description": "Parámetro necesario para la consulta en la tabla Centros de Acopio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        },
                        "examples": {
                            "integer": {
                                "summary": "Introduce el número de id de un centro de acopio.",
                                "value": 1
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso actualizado con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "No se ha encontrado el centro de acopio.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateCentroRequest": {
                "properties": {
                    "encargado_id": {
                        "description": "Encargado",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "estado_id": {
                        "description": "Estado",
                        "type": "integer",
                        "example": "7",
                        "nullable": false
                    },
                    "municipio_id": {
                        "description": "Municipio",
                        "type": "integer",
                        "example": "90",
                        "nullable": false
                    },
                    "description": {
                        "description": "Descripción del centro de acopio",
                        "type": "string",
                        "example": "Centro de Acopio de Valencia",
                        "nullable": true
                    },
                    "address": {
                        "description": "Dirección del centro de acopio",
                        "type": "string",
                        "format": "Valencia centro",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "UpdateCentroRequest": {
                "properties": {
                    "encargado_id": {
                        "description": "Encargado",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "estado_id": {
                        "description": "Estado",
                        "type": "integer",
                        "example": "7",
                        "nullable": false
                    },
                    "municipio_id": {
                        "description": "Municipio",
                        "type": "integer",
                        "example": "90",
                        "nullable": false
                    },
                    "description": {
                        "description": "Descripción del centro de acopio",
                        "type": "string",
                        "example": "Centro de Acopio de Valencia",
                        "nullable": true
                    },
                    "address": {
                        "description": "Dirección del centro de acopio",
                        "type": "string",
                        "format": "Valencia centro",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "sanctum",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}